hw-3 fix
