--- src/App.css
@@ -1,7 +1,0 @@
-.App {
-  background-image: url("./img/bg/bg.jpg");
-}
-body {
-  margin: 0;
-  padding: 0;
-}
