+++ src/components/Auth/index.js
@@ -1,9 +1,11 @@
 import React from 'react';
 import PropTypes from "prop-types";
+import { Logo } from 'loft-taxi-mui-theme';
 import { makeStyles } from '@material-ui/core/styles';
 import Grid from '@material-ui/core/Grid';
-import {LoginForm} from './LoginForm';
-import { Logo } from 'loft-taxi-mui-theme';
+import { LoginForm } from '../Login/LoginForm';
+import { SignupForm } from '../Signup/SignupForm';
+
 
 const BackgroundURL = '../background.jpg';
 
@@ -20,27 +22,27 @@ const useStyles = makeStyles(theme => ({
   }
 }));
 
-export const Login = () => {
+export const Auth = ({ isAuth }) => {
   const classes = useStyles();
 
   return (
     <div className={classes.login}>
-       <Grid 
+      <Grid
         container
         direction="row"
         justify="center"
         alignItems="center">
-          <Grid item xs={3}>
-            <Logo white animated />
-          </Grid>
-          <Grid item xs={3}>
-            <LoginForm/>
-          </Grid>
-       </Grid>
+        <Grid item xs={3}>
+          <Logo white animated />
+        </Grid>
+        <Grid item xs={3}>
+          {isAuth ? <LoginForm /> : <SignupForm />}
+        </Grid>
+      </Grid>
     </div>
   );
 }
 
-Login.propTypes = {
-  setRoute: PropTypes.func
+Auth.propTypes = {
+  isAuth: PropTypes.bool
 };
