+++ src/components/Login/LoginForm.js
@@ -1,13 +1,13 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
+import { shallowEqual, useSelector, useDispatch } from 'react-redux';
 import { makeStyles } from '@material-ui/core/styles';
 import Card from '@material-ui/core/Card';
 import Typography from '@material-ui/core/Typography';
 import Grid from '@material-ui/core/Grid';
 import TextField from '@material-ui/core/TextField';
 import Button from '@material-ui/core/Button';
-import { getAuth, fetchAuthRequest } from '../../modules/main';
-import {shallowEqual, useSelector, useDispatch} from 'react-redux';
+import { getAuth, fetchAuthRequest } from '../../modules/auth';
 import history from '../../history';
 
 const useStyles = makeStyles(theme => ({
@@ -47,45 +47,42 @@ export const LoginForm = () => {
     dispatch({
       ...authAction,
       payload: {
-          email: e.target.email.value,
-          password: e.target.password.value
+        email: e.target.email.value,
+        password: e.target.password.value
       }
     });
-
-    // console.log(auth);
-    // console.log(authAction);
-   }
+  }
 
   return (
     <Card className={classes.card}>
-            <form  onSubmit={onSubmit}>
-              <Grid container>
-                <Grid item xs={12}>
-                  <Typography variant='h4' component='h1'>
-                    Войти
+      <form onSubmit={onSubmit}>
+        <Grid container>
+          <Grid item xs={12}>
+            <Typography variant='h4' component='h1'>
+              Войти
                   </Typography>
-                </Grid>
-                <Grid item xs={12}>
-                  <p>
-                    Новый пользователь?{' '}
-                    <Link to={'/signup'} >
-                      Зарегистрируйтесь
+          </Grid>
+          <Grid item xs={12}>
+            <p>
+              Новый пользователь?{' '}
+              <Link to={'/signup'} >
+                Зарегистрируйтесь
                     </Link>
-                  </p>
-                </Grid>
-                <Grid item xs={12}>
-                  <TextField fullWidth required name="email" type="email" label="Имя пользователя"/>
-                </Grid>
-                <Grid item xs={12} >
-                  <TextField fullWidth required name="password" type="password" label="Пароль" margin="normal"/>
-                </Grid>
-                <Grid item xs={12} align="right" className={classes.btnGrid}>
-                  <Button type="submit" variant="contained" color="primary">
-                    Войти
+            </p>
+          </Grid>
+          <Grid item xs={12}>
+            <TextField fullWidth required name="email" type="email" label="Имя пользователя" />
+          </Grid>
+          <Grid item xs={12} >
+            <TextField fullWidth required name="password" type="password" label="Пароль" margin="normal" />
+          </Grid>
+          <Grid item xs={12} align="right" className={classes.btnGrid}>
+            <Button type="submit" variant="contained" color="primary">
+              Войти
                   </Button>
-                </Grid>
-              </Grid>
-            </form>
+          </Grid>
+        </Grid>
+      </form>
     </Card>
   )
 }
