+++ src/components/Map/index.js
@@ -1,8 +1,9 @@
-import React, {Component} from 'react';
-import { withStyles } from '@material-ui/styles';
+import React, { Component } from 'react';
 import mapboxgl from 'mapbox-gl';
+import { withStyles } from '@material-ui/styles';
 
-mapboxgl.accessToken = 'pk.eyJ1IjoidDB4YSIsImEiOiJjazV5MXJtNmoxM2l5M3BwbzcyYWowYXM3In0.HwMcjS3exgan06ermvYewg';
+
+mapboxgl.accessToken = 'pk.eyJ1IjoiY2hvdGluZWMiLCJhIjoiY2s1dXIxbDEyMDNqazNybGwzcTBydmdybyJ9.E0ZzR-BquMw7y5WGatf6XQ';
 
 const styles = theme => ({
   mapContainer: {
@@ -45,37 +46,38 @@ class MapComponent extends Component {
 
       let labelLayerId
       for (let i = 0; i < layers.length; i++) {
-          if (layers[i].type === 'symbol' && layers[i].layout['text-field']) {
-              labelLayerId = layers[i].id
-              break
-          }
+        if (layers[i].type === 'symbol' && layers[i].layout['text-field']) {
+          labelLayerId = layers[i].id
+          break
+        }
       }
 
       map.addLayer({
-          'id': '3d-buildings',
-          'source': 'composite',
-          'source-layer': 'building',
-          'filter': ['==', 'extrude', 'true'],
-          'type': 'fill-extrusion',
-          'minzoom': 15,
-          'paint': {
-              'fill-extrusion-color': '#aaa',
+        'id': '3d-buildings',
+        'source': 'composite',
+        'source-layer': 'building',
+        'filter': ['==', 'extrude', 'true'],
+        'type': 'fill-extrusion',
+        'minzoom': 15,
+        'paint': {
+          'fill-extrusion-color': '#aaa',
 
-              // use an 'interpolate' expression to add a smooth transition effect to the
-              // buildings as the user zooms in
-              'fill-extrusion-height': [
-                  'interpolate', ['linear'], ['zoom'],
-                  15, 0,
-                  15.05, ['get', 'height']
-              ],
-              'fill-extrusion-base': [
-                  'interpolate', ['linear'], ['zoom'],
-                  15, 0,
-                  15.05, ['get', 'min_height']
-              ],
-              'fill-extrusion-opacity': .6
-          }
-      }, labelLayerId)})
+          // use an 'interpolate' expression to add a smooth transition effect to the
+          // buildings as the user zooms in
+          'fill-extrusion-height': [
+            'interpolate', ['linear'], ['zoom'],
+            15, 0,
+            15.05, ['get', 'height']
+          ],
+          'fill-extrusion-base': [
+            'interpolate', ['linear'], ['zoom'],
+            15, 0,
+            15.05, ['get', 'min_height']
+          ],
+          'fill-extrusion-opacity': .6
+        }
+      }, labelLayerId)
+    })
   }
 
   render() {
@@ -83,7 +85,7 @@ class MapComponent extends Component {
 
     return (
       <div className={classes.mapContainer}>
-        <div ref={this.mapRef} className={classes.map}/>
+        <div ref={this.mapRef} className={classes.map} />
       </div>
     );
   }
