+++ src/components/Signup/SignupForm.js
@@ -1,13 +1,13 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
+import { shallowEqual, useSelector, useDispatch } from 'react-redux';
 import { makeStyles } from '@material-ui/core/styles';
 import Card from '@material-ui/core/Card';
 import Typography from '@material-ui/core/Typography';
 import Grid from '@material-ui/core/Grid';
 import TextField from '@material-ui/core/TextField';
 import Button from '@material-ui/core/Button';
-import { getRegister, fetchRegisterRequest } from '../../modules/main';
-import {shallowEqual, useSelector, useDispatch} from 'react-redux';
+import { getRegister, fetchRegisterRequest } from '../../modules/register';
 import history from '../../history';
 
 
@@ -35,7 +35,7 @@ export const SignupForm = () => {
   const registerAction = useSelector(fetchRegisterRequest, shallowEqual);
 
   const dispatch = useDispatch();
-  
+
   const onSubmit = (event) => {
     event.preventDefault();
 
@@ -60,7 +60,7 @@ export const SignupForm = () => {
 
   return (
     <Card className={classes.card}>
-      <form  onSubmit={onSubmit}>
+      <form onSubmit={onSubmit}>
         <Grid container>
           <Grid item xs={12}>
             <Typography variant='h4' component='h1'>
@@ -76,18 +76,18 @@ export const SignupForm = () => {
             </p>
           </Grid>
           <Grid item xs={12}>
-            <TextField fullWidth required name="email" type="email" label="Адрес электронной почты"/>
+            <TextField fullWidth required name="email" type="email" label="Адрес электронной почты" />
           </Grid>
           <Grid container spacing={2}>
             <Grid item xs={6} >
-              <TextField fullWidth required name="name" type="text" label="Имя" margin="normal"/>
+              <TextField fullWidth required name="name" type="text" label="Имя" margin="normal" />
             </Grid>
             <Grid item xs={6} >
-              <TextField fullWidth required name="scondName" type="text" label="Фамилия" margin="normal"/>
+              <TextField fullWidth required name="scondName" type="text" label="Фамилия" margin="normal" />
             </Grid>
           </Grid>
           <Grid item xs={12} >
-            <TextField fullWidth required name="password" type="password" label="Пароль" margin="normal"/>
+            <TextField fullWidth required name="password" type="password" label="Пароль" margin="normal" />
           </Grid>
           <Grid item xs={12} align="right" className={classes.btnGrid}>
             <Button type="submit" variant="contained" color="primary" data-testid="buttonLogin">
@@ -95,7 +95,7 @@ export const SignupForm = () => {
             </Button>
           </Grid>
         </Grid>
-        
+
       </form>
     </Card>
   );
