+++ src/modules/card/index.js
@@ -1,4 +1,4 @@
 export { default } from './reducer';
 export * from './selectors';
 export * from './actions';
-export * from './middlewares';
+export * from './cardMiddlewares';
