+++ src/modules/register/index.js
@@ -1,0 +1,4 @@
+export { default } from './reducer';
+export * from './selectors';
+export * from './actions';
+export * from './registerMiddlewares';
