+++ src/shared/Header/Header.js
@@ -7,8 +7,8 @@ import Toolbar from '@material-ui/core/Toolbar';
 import Typography from '@material-ui/core/Typography';
 import { Logo } from 'loft-taxi-mui-theme';
 import Button from '@material-ui/core/Button';
-import { fetchAuthRequest } from '../../modules/main';
-import {shallowEqual, useSelector, useDispatch} from 'react-redux';
+import { fetchAuthRequest } from '../../modules/auth';
+import { shallowEqual, useSelector, useDispatch } from 'react-redux';
 
 const useStyles = makeStyles(theme => ({
   root: {
@@ -40,24 +40,24 @@ export const Header = () => {
       }
     });
   }
-  
+
   return (
     <AppBar position="static" color="inherit">
-        <Toolbar>
-          <Typography variant="h6" color='inherit' className={classes.grow}>
-            <Logo/>
-          </Typography>
-          <Button color="inherit">
-            <Link to="/map" className={classes.headerLink}>Карта</Link>
-          </Button>
-          <Button color="inherit">
-            <Link to="/profile" className={classes.headerLink}>Профиль</Link>
-          </Button>
-          <Button id={'LogoutButton'} color="inherit" onClick={onClick}>
-            <Link to="/login" className={classes.headerLink}>Выйти</Link>
-          </Button>
-        </Toolbar>
-      </AppBar>
+      <Toolbar>
+        <Typography variant="h6" color='inherit' className={classes.grow}>
+          <Logo />
+        </Typography>
+        <Button color="inherit">
+          <Link to="/map" className={classes.headerLink}>Карта</Link>
+        </Button>
+        <Button color="inherit">
+          <Link to="/profile" className={classes.headerLink}>Профиль</Link>
+        </Button>
+        <Button id={'LogoutButton'} color="inherit" onClick={onClick}>
+          <Link to="/login" className={classes.headerLink}>Выйти</Link>
+        </Button>
+      </Toolbar>
+    </AppBar>
   );
 }
